Reads in all relevant data files from Presentation Data
Fixed a bug where MorningMemoryData has extra commas at the end of the line. Function now detects when there are extra commas at the end of the line, reads the file in with readLines, removes the extra commas, then converts to a dataframe before binding to the existing data frame.

Note: function does not currently read in the native Presentation output file which is poorly formatted

Note: there is a bug with reading in the Cues data which will need to be addressed