Added reading in and creating a 'usable' dataframe of the PresentationData raw output files to the ReadPresentationData.R script
Needed to create a parsing function in order to properly read the messy data.

Currently puts all output files (evening, morning, and overnight) into one large file.

Note that 8002 E1 was misnamed 8003 (forgot to change when running it). Still need to verify that I have the correct file identified in the script exception.