corrected AttemptData read-in: was converting the 9,7,N/A to 1,2,3 because it was trying to force it into a number (I think). Now shifted the columns for 8001 appropriately so fixed that problem.
Also removed the unnecessary "X" column which was created because of an extra column in the header of the attemptdata file I wrote from Presentation. That is now fixed in the Presentation script as well as changing 9's and 7's to spaces and 2's etc. as the button name instead of the button number.